@@ -12,11 +12,13 @@
         {"name": "Jonathan Wage", "email": "jonwage@gmail.com"}
     ],
     "require": {
-        "php": ">=5.3.2",
-        "doctrine/common": ">=2.4,<2.8-dev"
+        "php": "^7.1",
+        "ext-pdo": "*",
+        "doctrine/common": "^2.7.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.*",
+        "phpunit/phpunit": "^5.4.6",
+        "phpunit/phpunit-mock-objects": "!=3.2.4,!=3.2.5",
         "symfony/console": "2.*||^3.0"
     },
     "suggest": {
@@ -26,9 +28,12 @@
     "autoload": {
         "psr-0": { "Doctrine\\DBAL\\": "lib/" }
     },
+    "autoload-dev": {
+        "psr-0": { "Doctrine\\Tests\\": "tests/" }
+    },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.5.x-dev"
+            "dev-master": "2.6.x-dev"
         }
     },
     "archive": {
