@@ -10,14 +10,14 @@
         }
     ],
     "require": {
-        "php":                 "~7.0",
-        "composer-plugin-api": "^1.0"
+        "php":                 "^7.1.0",
+        "composer-plugin-api": "^1.0.0"
     },
     "require-dev": {
-        "phpunit/phpunit":   "^6.4",
-        "humbug/humbug":     "dev-master",
-        "composer/composer": "^1.3",
-        "ext-zip":           "*"
+        "phpunit/phpunit":     "^7.0.0",
+        "infection/infection": "^0.7.1",
+        "composer/composer":   "^1.6.3",
+        "ext-zip":             "*"
     },
     "autoload": {
         "psr-4": {
