@@ -19,16 +19,20 @@
         }
     ],
     "require": {
-        "php":                       "7.0.0 - 7.0.5 || ^7.0.7",
-        "zendframework/zend-code":   "3.0.0 - 3.0.2 || ^3.0.4",
-        "ocramius/package-versions": "^1.0"
+        "php":                       "^7.1.0",
+        "zendframework/zend-code":   "^3.1.0",
+        "ocramius/package-versions": "^1.1.1"
     },
     "require-dev": {
-        "ext-phar":                    "*",
-        "phpunit/phpunit":             "^5.4.6",
-        "squizlabs/php_codesniffer":   "^2.6.0",
-        "couscous/couscous":           "^1.4.0",
-        "phpbench/phpbench":            "^0.11.2"
+        "ext-phar":                     "*",
+        "phpunit/phpunit":              "^5.6.4",
+        "phpunit/phpunit-mock-objects": "^3.4.1",
+        "squizlabs/php_codesniffer":    "^2.7.0",
+        "couscous/couscous":            "^1.5.2",
+        "phpbench/phpbench":            "^0.12.2",
+        "phpstan/phpstan":              "^0.6.4",
+        "nikic/php-parser":             "^3.0.4",
+        "humbug/humbug":                "dev-master@DEV"
     },
     "suggest": {
         "ocramius/generated-hydrator": "To have very fast object to array to object conversion for ghost objects",
