@@ -12,14 +12,17 @@
             "homepage": "https://sonata-project.org/"
         }
     ],
+    "require": {
+        "php": "^7.1",
+        "psr/log": "^1.0"
+    },
     "require-dev": {
-        "psr/log": "~1.0",
-        "doctrine/orm": "~2.2",
-        "predis/predis": "~0.8,<1.0",
-        "doctrine/phpcr-odm": "~1.0",
-        "jackalope/jackalope-doctrine-dbal": "~1.0",
-        "symfony/phpunit-bridge": "~2.7|~3.0",
-        "fabpot/php-cs-fixer": "~0.1|~1.0"
+        "doctrine/orm": "^2.2",
+        "predis/predis": "^0.8",
+        "doctrine/phpcr-odm": "^1.0",
+        "jackalope/jackalope-doctrine-dbal": "^1.0",
+        "symfony/phpunit-bridge": "^3.3",
+        "sllh/php-cs-fixer-styleci-bridge": "^2.0"
     },
     "suggest": {
         "doctrine/orm": "ORM support",
@@ -28,15 +31,15 @@
         "ext-memcached": "Caching with ext/memcached",
         "predis/predis": "Install redis php"
     },
+    "autoload": {
+        "psr-4": { "Sonata\\Cache\\": "src/" }
+    },
+    "autoload-dev": {
+        "psr-4": { "Sonata\\Cache\\Tests\\": "tests/" }
+    },
     "extra": {
         "branch-alias": {
             "dev-master": "1.x-dev"
-        }
-    },
-    "autoload": {
-        "psr-4": {
-            "Sonata\\Cache\\Tests\\": "test/",
-            "Sonata\\Cache\\": "lib/"
         }
     }
 }
