@@ -167,8 +167,10 @@ class CRUDController extends Controller
     }
 
     /**
-     * @param $id
-     */
+    *Ajout tout les utilisateurs a la newsletter
+    *@param $id
+    *
+    **/
     public function addUserListAction($id)
     {
     	$em = $this->getDoctrine()->getManager();
@@ -183,8 +185,9 @@ class CRUDController extends Controller
 
         return new RedirectResponse($this->admin->generateUrl('list'));
     }
-        /**
-     * @param $id
+     /**
+     * Ajout tout les utilisateurs de la promo a la newsletters
+     * @param $id, $promotion
      */
     public function addUserPromoAction($id,$promotion)
     {
